Renamed TodoTemplate.Infra.Azure to TodoTemplate.Infra
