+++ src/Tooling/Bit.Tooling.Templates/TodoTemplate/Infra/Pulumi.yaml
@@ -1,3 +1,3 @@
-﻿name: TodoTemplate.Infra.Azure
+﻿name: TodoTemplate.Infra
 runtime: dotnet
 description: TodoTemplate infra as a code
